Revert "Do not autostart the UI from the daemon on Windows"
This is here because the kdeconnect shortcut launches the
app, not the systray. The app launches the daemon via dbus
but without this, we don't have a systray.

Ideally the systray and daemon would be the same process.

Also ideally we would have a way to launch the daemon and
systray at boot, so people don't have to launch the app
just for the side effect of starting the daemon.

This reverts commit 1dcfaf5108d4b96db1162fcc95990eae615e16f6.