Fix SMS app icons and thumbnails
## Summary

Due to the difficult-to-test Qt5 -> Qt6 transition, there were some GUI errors with the SMS app:
- Contact photos were missing
- Attachment previews, if present, were in the place where the contact photo should be

This also takes a shot at fixing the long-standing issue that attachment previews were shown much taller than the row item, drawing over the items above and below. 

## Test Plan

### Before:
As in description, the conversations list items were not correct.

![image](/uploads/f68b662fecd6a4826986ede6e8191470/image.png)

### After:
Contact photos are shown to the left of the text preview, attachment preview, if present, is shown to the far right.

![image](/uploads/95f2b4d6e6ff26371a2f36d97fc3f52b/image.png)