kcm: improve description field in json
The kcm json currently lists the description as "Connect and sync your
devices". This is hard to distinguish from the KDE Connect application
at a glance, and does not follow the general pattern in kcm descriptions
to use either a configuration-based verb or noun (e.g. configure, edit,
settings).

This change sets the description to "Configure device synchronization".

BUG: 466256