plasmoid: improve description field in json
The plasmoid currently has the decription "Show notifications from your
devices using KDE Connect. This is wordy and misleading as
the plasmoid does much more than display notifications; it's used to
show the  devices and their current state (e.g. battery) and provides
access to some of the broader KDE Connect functionality, such as the
find device function or file sharing.

This change sets the description to "Manage connected devices".