settings: improve .desktop GenericName and Comment
The .desktop file for the non-Plasma settings app currently has a
sentence as the GenericName, which is supposed to be the term of a
class of applications, and does not have a Comment field, which is
suitable for descriptions in the form of a sentence.

This change sets the GenericName to "Device Synchronization Settings",
and adds a Domment field that matches the description for the kcm.