smsapp: improve .desktop GenericName and Description
The KDE Connect SMS app currently has the GenericName "SMS" which is the
name of a protocol, not the name for a class of applications, and the
Comment "Text Messaging", which is correct but very short for a field
that is intended to give a short but more exhaustive description of an
application's functions.

This change sets the GenericName To "Text Messaging" and the Comment to
"Read and send SMS messages on connected devices".