app: improve screen reader a11y
There are several instances in the KDE Connect app where information is
not properly presented to screen reader users. This includes ListViews,
which do not make use of orca's function to announce that this is a
(keyboard-navigable) list and how many items it contains, buttons that do
not have informative labels, and extra information that is presented
visually but not read out, such as the connection status in the device
list (e.g. "Device disconnected") or the label informing the user how to
exit a locked Remote Control plugin.

This change sets various Accessible properties to improve this, and uses
the new announce method in Qt 6.8 (if available) to read out the
conditional label in Remote Control.