app: close page when Escape pressed
kdeconnect-app currently mostly does nothing when Escape is pressed.
This is the conventional key for abort, so users likely intend to abort
the plugin or settings page that they are currently on. (The global
settings dialog is already closed on Escape) This is inconvenient for
sighted users, and can lead to confusion for screen reader users.

This change binds Escape to closing the currently active page unless it
is the device page, allowing users to return there quickly.