app/Main: set height for device list ScrollView
The device list in the sidebar uses a ScrollView in case the list of
devices is too long to display completely. However, it does not have a
height set, so the ScrollView uses the child size as the default, and
content is pushed outside the window area (and left inaccessible) rather
than introduce a scrollable area.

This change sets the ScrollView size through Layout.fillHeight, enabling
it to fulfill its intended purpose of scrolling if the device list gets
large.