app: Eliminate some runtime warnings
Moved from !769 as that one is becoming a bit unwieldy and they are not really related.

----

Qt has deprecated injected signal handlers in favor of explicit formal
parameters. mousepad still has some uses of injection, leading to
unnecessary runtime warnings.

This change swaps these for formal parameters.

----

PluginItem initalizes all plugin pages by setting the device property.
Many pages do not make use of this, and do not have the property
declared. This is harmless, but leads to an unnecessary runtime warning.

This change declares the property for all pages, eliminating these
warnings.