app/DevicePage: use sections to structure list
The DevicePage lists the available options in an unstructured way. One
way to structure them is by whether they perform an instantly completed
action (like Find My Phone or Lock Screen), or open a subpage that
allows several further options, like Media Remote or Remote Input. This
is a relevant distinction in many ways; for example, the second group
can be opened without notable side effects, but clicking on members of
the second generally has some, such as leaking the clipboard contents to
the phone or causing it to ring loudly.

This change splits the list into "Actions" and "Controls" sections
according to this splits. The order of list entries is largely kept the
same, except for moving some to make the categories contiguous.