DevicePage: add accelerators to plugin ListView
The device page has a list containing actions and controls that are
available for this device. The can be navigated with the arrow keys, but
there are no accelerators/mnemonics for direct access even though the
list is rather short, with fewer than ten entries at most. Accelerator
keys are helpful for keyboard users, making them an a11y-relevant
feature.

This change adds accelerator support. No default accelerators are
suggested, relying on automatic assignment to make things easier to
manage for developers and translators. Making this change requires
moving the filtering of active plugins into the view rather using
filter() on the list, as the mnemonic assignment is executed for the
resulting list each time a plugin gets loaded, and the keys assigned
in each of those runs are not available in subsequent runs, causing the
final list to have no mnemonics assigned.