app/Main: add accelerators to global drawer
The list of devices in the sidebar does not have accelerator keys/
mnemonics set. The number of devices should generally be small enough
that accelerators would make sense to allow users to switch between
devices with a single keystroke. In particular, this makes navigating
the app more convenient for keyboard (and thus screen reader) users.

This change adds accelerator keys for the sidebar items. The mnemonics
are automatically assigned from the device names.