app: set default focus element for plugins
Most plugin pages and plugin settings pages have no element that has
default focus. As a consequence, when the page is first opened, focus
ends up on a non-interactive element and keyboard users need to press
tab first to interact with anything.

This change sets an item to receive focus by default, so keyboard users
can start interacting with the page immediately.