PluginSettings: improve ListView keyboard navigation
The plugin list delegate on the plugin settings page has a checkbox that
allows the user to enable or disable the plugin. This UI element is
somewhat awkward to use for keyboard users, as the checkboxes all appear
in the tab chain, and are not linked to the currently selected item; to
toggle the box for an entry, the user will have to hit the tab key for
each element above it in the list, even if the item is already selected.
There are also screen reader a11y issues. The delegate itself does
nothing; the settings for that particular plugin are accessible through
an action button on the delegate.

This changes removes the checkboxes from the tab chain; acting on the
delegate itself now toggles the box. The delegate identifies as a
checkbox to inform screen reader users about this and the current state.