app: allow more arrow key navigation
The various ListViews (inclusing e.g. in the global drawer or the device
page) can already be navigated with the arrow keys, but the plugin pages
and plugin settings pages that are not solely list views cannot. This
makes keyboard navigation overly reliant on the tab key. Certain parts
of the sidebar, such as the search field, are also not reachable

This change implements arrow key navigation on the plugin and plugin
settings pages, as well as in the sidebar.

(cherry picked from commit 85ef43ae9fe88de2f81df0adef9710303508dcde)