Register ourselves to xdg-desktop-portal
Authorization in xdg-desktop-portal, for example for RemoteDesktop
sessions, is based on the app id.

xdg-desktop-portal derives the app id from the cgroup the caller is
in. For us this is problematic because depending on how kdeconnectd
was launched (autostart, DBus activation, or manual start) it can
yield different results.

To address this make use of new portal API to explicitly set the ID

BUG: 480412