interfaces: make canAddCommand as CONSTANT for dbus
qdbusxml2cpp does not carry over that a property over dbus is marked as
CONSTANT. This causes qml code depending on these properties to complain
about missing NOTIFY on these properties. In the remoteCommands dbus
interface, this is already handled for the deviceId property, but not
for the canAddCommand property. As a result, there is a warning printed
for each device on plasmoid startup.

This change sets the property explicitly to CONSTANT for the dbus
interface, resolving these warnings.