plugins/findthisdevice: Fix issues surrounding selecting a custom sound
The plugin configuration page for Find This Device uses the pathExists
method of the FindThisDeviceHelper singleton to determine whether to
enable the play button. This method is not marked as Q_INVOKABLE and
so the method call from QML fails with "Property 'pathExists' is not a
function".

This change marks the method as Q_INVOKABLE so that it is actually
exported to QML.

----



The sound file selection sets the currently chosen sound file as the
current file for the file picker dialog that allows users to choose
which sound to play. Dialogs.FileDialog expects this to be provided as
an URL (absolute or relative), but the file is provided as a path, so
without the required file:// prefix. As a consequence, an error is
printed, and the dialog is not opened in the correct folder nor with
the correct item pre-selected.

This change adds the file:// prefix to the path. The deprecated
currentFile property is also replaced with the recommended selectedFile
property.