app/DevicePage: only show section headers with 2+ sections
The plugin list hides delegates instead of pre-filtering the list, as
otherwise the mnemonic assignment to list entries would fail due to the
mnemonics being recomputed as the plugins are asynchronously loaded and
the bindings reloaded each time. This works well in the regular case,
but will cause headers for empty sections to be printed. Unfortunately,
the same hiding mechanisms for delegates will not work for section
headers, ListView will always display them.

Instead, we can just show the section headers only if delegates for at
least two sections are visible – there are only two sections, and in the
edge case of only plugins of one type being enabled there's not really
a point to the sections in the first place, so not showing headers then
would also make sense.