app/DevicePage: use DelegateModel to filter list
Previously, the list of plugins was constantly refiltered usning
filter() when a new  plugin was loaded. The introduction of mnemonic
support for the entries in the list, as the new list conflicted with the
mnemonics already assigned. Unfortunately, the strategy of just hiding
the delegate for unloaded plugins breaks many other things: for the
section headers a workaround was already added, but it also has
implications for keyboard navigation, the reporting of the number of
list entries for accessibility support, and possibly others.

Instead of adding more and more hacks to work around this, we switch
back to a strategy that actually filters the model, but this time one
that keeps the objects the same and only chenges whether they are in
the model or not. QML provides a type for this purpose, even if it
is a little cumbersome to use, namely DelegateModel. This also allows us
to remove the visibility/item height and section display workarounds.