plugins/mpriscontrol: port away from deprecated signal
QDBusConnectionInterface::serviceOwnerChanged is deprecated and should
be replaced with QDBusServiceWatcher. The plugin already creates the
corresponding object, so we can just connect to that instead. We can
also filter the ServiceWatcher on the mpris namespace; this was removed
in the past but seems to work now and is done this way in Plasma's
libkmpris.