Fix build on platforms with sizeof(qsizetype) < sizeof(qint64)
## Summary

`state->read_buffer.size()` returns `qsizetype` which is defined as `ssize_t`. On some platforms, this is less than the size of `qint64`.

This change casts the return of `state->read_buffer.size()` into `qint64` before comparing it to the `maxlen` parameter in `MultiplexChannel::readData`.

Includes defensive assert for the possible, likely distant, future in which `qint64` is not enough to contain a `ssize_t` value, in which case I cannot guarantee the rest of the code will perform correctly.

## Test Plan

### Before:
Build fails on 32-bit Debian 13 and Ubuntu 25.04 armhf package infra.

```
/core/backends/bluetooth/multiplexchannel.cpp:55:42: error: no matching function for call to ‘min(qint64&, qsizetype)’
   55 |         const auto num_to_read = std::min(maxlen, state->read_buffer.size());
      |
<snip>
/usr/include/c++/14/bits/stl_algobase.h:233:5: note: candidate: ‘template<class _Tp> constexpr const _Tp& std::min(const _Tp&, const _Tp&)’
  233 |     min(const _Tp& __a, const _Tp& __b)
      |     ^~~
/usr/include/c++/14/bits/stl_algobase.h:233:5: note:   template argument deduction/substitution failed:
/<<PKGBUILDDIR>>/core/backends/bluetooth/multiplexchannel.cpp:55:42: note:   deduced conflicting types for parameter ‘const _Tp’ (‘long long int’ and ‘qsizetype’ {aka ‘int’})
   55 |         const auto num_to_read = std::min(maxlen, state->read_buffer.size());
```

### After:
Build passes on my local 32-bit Debian 13 and Ubuntu armhf.