WIP removing KCM
TODO: Make the app open a specific device if --device <ID> is passed
TODO: Make the app open a specific plugin config if --plugin-config <ID>
is passed together with --device
TODO: Make the plasmoid open the app instead of trying to open the KCM
TODO: Maybe add a placeholder KCM that sends people to the app? Maybe not?