Bluetooth provider - make non-blocking
Investigating https://bugs.kde.org/show_bug.cgi?id=504690 and https://bugs.kde.org/show_bug.cgi?id=503431
the "hanging problem" from last year wasn't solved, it was only hidden by moving the blocking to
inside kdeconnectd rather than all the way out to the plasmoid. At the time I didn't think this
mattered, but given that users can and do notice that kdeconnectd isn't responsive during that
period; it's not really a good enough fix.

This puts the construction of the provider into a worker thread and implement a messagepump.
This is working for me in both the no-hardware (i.e. there is no delays within Kdeconnect) and
the hardware (I am able to pair and use bluetooth etc).