smsapp: Restore the last used device on start-up
I usually have two Android devices connected to my computers: a
non-cellular tablet and my actually-SMS-capable phone. Due to some
sorting somewhere, my tablet is always the first one selected when
starting KDE Connect SMS and thus none of the messages I want to see are
shown.

So I made it so AppData (the data object holding among other things, the
currently used device) writes and loads the last used device into
KStateConfig. This works really well, except with one caveat: if the
device is no longer reachable or is otherwise invalid, the user has to
manually reselect a valid device. I think this is overall pretty minor,
as this is the rare case (no one is frequently creating invalid
devices.) Trying to work around this will be hard due to how the device
list is initialized in our app too.