smsapp: Add download button for attachments
This also adds a helper function to open the save dialog from QML, and introduces a helper C++ singleton for attachments, as there is no way, as far as I know, to save a file from QML alone.

Note that this isn't really "saving" anything, as the file is already on disk in \~/.cache, but most users wouldn't know how to retrieve an attachment from there, so "saving" seems to be a more appropriate terminology.

EDIT: @sredman suggested "downloading", which makes sense to me, since from the point of view of the user, it is bringing a file from their phone/remote device to their local one!

I also added a small patch to fix the background colour of the attachment viewer being white in dark mode, which I did while working on this. LMK if I should make a separate MR for this!